move behavior test
